fix: move /hooks to /src
